+++ docker-compose.yml
@@ -1,7 +1,7 @@
-version: '3'
+version: '3.8'
 services:
-  mysql:
-    image: mariadb:11.6.2
+  db:
+    image: mysql:8.0
     ports:
       - "3306:3306"
     environment:
@@ -13,7 +13,7 @@ services:
       - ./data/mysql/:/var/lib/mysql
 
   redis:
-    image: redis:7.4.2
+    image: redis:latest
     container_name: redis
     ports:
       - "6379:6379"
