+++ src/test/java/com/server/concert_reservation/ConcertApplicationTests.java
@@ -1,15 +1,16 @@
 package com.server.concert_reservation;
 
+import com.server.concert_reservation.config.RdbTestcontainersConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 
-@Import(TestcontainersConfiguration.class)
+@Import(RdbTestcontainersConfig.class)
 @SpringBootTest
 class ConcertApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
