+++ src/test/java/com/server/concert_reservation/TestConcertApplication.java
@@ -1,11 +1,12 @@
 package com.server.concert_reservation;
 
+import com.server.concert_reservation.config.RdbTestcontainersConfig;
 import org.springframework.boot.SpringApplication;
 
 public class TestConcertApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.from(ConcertApplication::main).with(TestcontainersConfiguration.class).run(args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.from(ConcertApplication::main).with(RdbTestcontainersConfig.class).run(args);
+    }
 
 }
