+++ src/test/java/com/server/concert_reservation/config/RdbTestcontainersConfig.java
@@ -1,0 +1,35 @@
+package com.server.concert_reservation.config;
+
+import jakarta.annotation.PreDestroy;
+import org.springframework.context.annotation.Configuration;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.utility.DockerImageName;
+
+@Configuration
+public
+class RdbTestcontainersConfig {
+
+    @Container
+    static MySQLContainer<?> MYSQL_CONTAINER;
+
+    static {
+        MYSQL_CONTAINER = new MySQLContainer<>(DockerImageName.parse("mysql:8.0"))
+                .withDatabaseName("concert")
+                .withUsername("test")
+                .withPassword("test");
+        MYSQL_CONTAINER.start();
+
+        System.setProperty("spring.datasource.url", MYSQL_CONTAINER.getJdbcUrl() + "?characterEncoding=UTF-8&serverTimezone=UTC");
+        System.setProperty("spring.datasource.username", MYSQL_CONTAINER.getUsername());
+        System.setProperty("spring.datasource.password", MYSQL_CONTAINER.getPassword());
+    }
+
+    @PreDestroy
+    public void preDestroy() {
+        if (MYSQL_CONTAINER.isRunning()) {
+            MYSQL_CONTAINER.stop();
+        }
+    }
+
+}
